moved stage music to a shared music folder
Music should be a shared resource, since most stages will play music
from other stages. If music is shared, it will make resources much
smaller, and give the ability to read from a single folder to select
music. There can be a music property in some stage data, (not the main
stage.plist) that specifies different songs that are preferred for the
stage.