fixed StageLoader and CharacterLoader load functions
removed loadAssets and replaced with reloadAssets. The reloadAssets
function is for when a new theme is being applied, and the icons and
such must be reloaded. Actual stage assets, like the stage sprites and
music and such, will be loaded from the Stage module, and that load will
be facilitated by StageLoader.
StageLoader and CharacterLoader are there to load the main info about
the character and stage for the menu, and then load the actual assets
for a specific Stage or Character when told to (that way it doesn't take
up massive amounts of memory)