BADASS STRING CLASS
and some other changes
But seriously. This string class is fucking TEMPLATED. Damn son.
I will admit I don't like the way I implemented the string exceptions,
but cmon man I had to. They had a dependency on string that had to be
removed. I'll also fix "split" so that it returns an ArrayList
eventually. But later.